+++ src/net/game/aura/Aura.java
@@ -6,7 +6,7 @@ public class Aura {
 
 	private final int id;
 	private final int duration;
-	private final int defaultNumberStack;
+	private final byte defaultNumberStack;
 	private final AuraEffect effect1;
 	private final int effectValue1;
 	private final AuraEffect effect2;
@@ -24,7 +24,7 @@ public class Aura {
 	private final boolean isStackable;
 	private final boolean isMagical;
 	
-	public Aura(int id, String name, String sprite_id, int spellTriggeredOnFade, int duration, boolean isStackable, int defaultNumberStack, int tickRate, boolean lowDispellable, boolean highDispellable, AuraEffect effect1, int effectValue1, AuraEffect effect2, int effectValue2, AuraEffect effect3, int effectValue3, boolean isBuff, boolean isVisible, boolean isMagical) {
+	public Aura(int id, String name, String sprite_id, int spellTriggeredOnFade, int duration, boolean isStackable, byte defaultNumberStack, int tickRate, boolean lowDispellable, boolean highDispellable, AuraEffect effect1, int effectValue1, AuraEffect effect2, int effectValue2, AuraEffect effect3, int effectValue3, boolean isBuff, boolean isVisible, boolean isMagical) {
 		this.id = id;
 		this.name = name;
 		this.sprite_id = sprite_id;
@@ -71,7 +71,7 @@ public class Aura {
 		return this.duration;
 	}
 	
-	public int getDefaultNumberStack() {
+	public byte getDefaultNumberStack() {
 		return this.defaultNumberStack;
 	}
